requirements.txt
# --- Web Framework & Server ---
flask
python-dotenv

# --- Core AI/ML Libraries for Local Model ---
torch>=2.0.0
transformers==4.36.2 
peft>=0.7.1
bitsandbytes>=0.41.1
accelerate>=0.25.0
trl>=0.7.4
datasets>=2.15.0

# --- RAG (Retrieval-Augmented Generation) Framework ---
langchain
langchain-community
sentence-transformers
faiss-cpu

# --- Document Processing ---
PyPDF2
python-docx

# --- API Clients ---
groq

# --- Evaluation ---
rouge-score